+++ src/utilities/wrappers/RedBlackTreeWrapper.h
@@ -18,7 +18,7 @@ public:
 	void remove(int) override;
 	bool isValid() override;
 	int* getAllKeys(int& length) override;
-	int calculateSpaceComplexity() override;
+	long long calculateSpaceComplexity() override;
 	void destroy() override;
 };
